% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{add_citations_from_relationships}
\alias{add_citations_from_relationships}
\title{Title}
\usage{
add_citations_from_relationships(
  relationship_child,
  citation_type,
  citations_list,
  clean_identifier
)
}
\arguments{
\item{clean_identifier}{}
}
\value{

}
\description{
Title
}
