% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_identifier}
\alias{clean_identifier}
\title{Reformat DOI to a standard format}
\usage{
clean_identifier(identifier)
}
\arguments{
\item{identifier}{(character) DOI string}
}
\value{
(character) DOI string in standard format, e.g. 10.1029/2020gb006552. Without URL heads like "https://doi.org" or the "doi:" prefix.
}
\description{
Reformat DOI to a standard format
}
