% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_citation_table.R
\name{create_citation_table}
\alias{create_citation_table}
\title{Title}
\usage{
create_citation_table(scope, file = NULL, standalone_dois = NULL, meta = NULL)
}
\arguments{
\item{standalone_dois}{}
}
\value{

}
\description{
Title
}
