% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{get_citation_for_doi}
\alias{get_citation_for_doi}
\title{Query the CrossRef API to get the details of a publication}
\usage{
get_citation_for_doi(doi)
}
\arguments{
\item{doi}{(character) A publication's DOI}
}
\value{
(list) A list format with the citation, title, authors, and pubdate. These are empty strings if the API call is unsuccessful.
}
\description{
Query the CrossRef API to get the details of a publication
}
