% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{get_citations_for_doi}
\alias{get_citations_for_doi}
\title{Get DataCite citations for EDI data packages}
\usage{
get_citations_for_doi(doi, remove_preprints = FALSE)
}
\arguments{
\item{doi}{(character) One DOI string from EDI data package to lookup citations for}

\item{remove_preprints}{(logical) TRUE/FALSE on whether to remove published preprints. Defaults to FALSE.}
}
\value{
(list) Unnamed list of citations found in DataCite
}
\description{
Get DataCite citations for EDI data packages
}
