% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{get_meta_for_all_items_in_scope}
\alias{get_meta_for_all_items_in_scope}
\title{Returns a dictionary of metadata for all packages in a given scope.}
\usage{
get_meta_for_all_items_in_scope(scope)
}
\arguments{
\item{scope}{(character) EDI scope, e.g., knb-lter-ble}
}
\value{
(list) A dictionary of metadata for all packages in the given scope, indexed by package
}
\description{
Returns a dictionary of metadata for all packages in a given scope.
}
