% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{get_package_info}
\alias{get_package_info}
\title{Returns a dictionary of metadata for a given package revision.}
\usage{
get_package_info(scope, id, revision)
}
\arguments{
\item{scope}{(character) The scope of the package, e.g., knb-lter-ble}

\item{id}{(numeric) The package ID, e.g., 1}

\item{revision}{(numeric) The revision number, e.g., 1}
}
\value{
(list) A dictionary of metadata for the given package revision
}
\description{
Returns a dictionary of metadata for a given package revision.
}
