% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{add_citations_from_relationships}
\alias{add_citations_from_relationships}
\title{retrieve citations from datacite REST API}
\usage{
add_citations_from_relationships(
  relationship_child,
  citation_type,
  citations_list,
  clean_identifier
)
}
\arguments{
\item{relationship_child}{(list)}

\item{citation_type}{(character)}

\item{citations_list}{(list)}

\item{clean_identifier}{(character)}
}
\value{

}
\description{
retrieve citations from datacite REST API
}
