% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{get_citations_for_doi}
\alias{get_citations_for_doi}
\title{Query the DataCite API for publications that have cited a given data DOI}
\usage{
get_citations_for_doi(doi)
}
\arguments{
\item{doi}{(character) Data package DOI, taken from EDI for example}
}
\value{
(list) Unnamed list of citations
}
\description{
Query the DataCite API for publications that have cited a given data DOI
}
